[ELF] Add --archive-as-start-lib to parse archives as --start-lib object files
https://maskray.me/blog/2022-01-16-archives-and-start-lib

We intern symbol names in archive indexes and .o symbol tables. For every
definition in an extracted archive member, we intern the symbol twice, once for
the archive index, once for the .o symbol table after extraction. This is
inefficient.

Symbols in a --start-lib ObjFile/BitcodeFile are only interned once.