[ELF] Demote symbols in /DISCARD/ discarded sections to Undefined
When an input section is matched by /DISCARD/ in a linker script, GNU ld
reports errors for relocations referencing symbols in the section:

    `.aaa' referenced in section `.bbb' of a.o: defined in discarded section `.aaa' of a.o

Implement the error by demoting eligible symbols to `Undefined` and
changing STB_WEAK to STB_GLOBAL.

It's arguable whether a weak reference to a discarded symbol should lead to
errors. GNU ld reports an error and our demoting approach reports an error as
well.

Close #58891

Co-authored-by: Bevin Hansson <bevin.hansson@ericsson.com>