[AArch64,ELF] Omit initial $d of non-text sections
To remove redundant $d mapping symbols in non-text sections, set the
start state to `EMS_Data`. This works since instructions in non-text
sections are only expected in error scenarios.

This aligns with aaelf64, which specifies:

> A section that contains instructions must have a mapping symbol defined at the beginning of the section. If a section contains only data no mapping symbol is required.

This patch decreases the .symtab size significantly:

```
% bloaty a64-1/bin/clang -- a64-0/bin/clang
    FILE SIZE        VM SIZE
 --------------  --------------
  +0.0%     +16  +0.0%     +16    .text
  [ = ]       0  -1.8%     -16    .relro_padding
  -0.7%  -154Ki  [ = ]       0    .strtab
  -6.8%  -561Ki  [ = ]       0    .symtab
  -0.5%  -715Ki  [ = ]       0    TOTAL
```

---

If aaelf64 is revised to allow omission of the $x at the beginning of a
text section, we could emit an end $x if a text section ends with data.
This would ensure the combined text sections contain accurage data
within code information.

Pull Request: https://github.com/llvm/llvm-project/pull/99580