[AArch64,ELF] Omit initial $x of text sections
Introduce a cl::opt option "aarch64-initial-mapping-for-text" to allow
an alternative behavior:

The assembler no longer automatically adds an initial $x to a text
section. Instead, it adds an ending $x only if the final data is not
instructions. During the linking process, when sections are combined
into the final output, the subsequent section starts with the A64 state.

```
.section .text1,"ax"
nop
// emit $d
.long 42
// emit $x

.section .text2,"ax"
nop
```

When .text1 is assembled using the traditional behavior, the absence of
an initial $x symbol at .text2 (new behavior) might lead to the date
state for disassemblers. However, this mix-and-match isn't a major issue
for a significant portion of users.

Fix a bug that `getPreviousSection().first` instead of
`getCurrentSection().first` is used.